CHANGE term->text_angle(int ang) to term->text_angle(float ang)
This is a change in the terminal API.  All terminals must be modified
to accommodate it, unless of course they do not support text rotation.
Old terminals using null_text_angle or b_text_angle are probably OK
since they go through a shared routine that has been converted.

Not all terminals have been updated to actually use the floating
point value; some just convert it immediately to (int) and keep on
doing what they did before even though in principle they could
handle a non-integer rotation value.

These include:  canvas pdf pm tkcanvas win

+++ src/bitmap.c
@@ -1226,7 +1226,7 @@ b_put_text(unsigned int x0, unsigned int y0, const char *str)
 
 
 int
-b_text_angle(int ang)
+b_text_angle(float ang)
 {
     b_angle = ang;
     return TRUE;
